Refactor lib folder to organise functions by purpose
