+++ next/components/article-list-item.tsx
@@ -4,7 +4,8 @@ import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import classNames from "classnames";
 
-import { absoluteUrl, formatDate } from "@/lib/utils";
+import { absoluteUrl } from "@/lib/drupal/absolute-url";
+import { formatDate } from "@/lib/utils";
 import { ArticleTeaser } from "@/lib/zod/article-teaser";
 
 interface ArticleListItemProps {
