+++ next/components/article-teaser.tsx
@@ -3,7 +3,8 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 
-import { absoluteUrl, formatDate } from "@/lib/utils";
+import { absoluteUrl } from "@/lib/drupal/absolute-url";
+import { formatDate } from "@/lib/utils";
 import { ArticleTeaser } from "@/lib/zod/article-teaser";
 
 interface ArticleTeaserProps {
