+++ next/lib/drupal/absolute-url.ts
@@ -1,0 +1,5 @@
+import { env } from "@/env";
+
+export function absoluteUrl(input: string) {
+  return `${env.NEXT_PUBLIC_DRUPAL_BASE_URL}${input}`;
+}
