+++ next/pages/api/register.ts
@@ -1,5 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next";
-import { drupal } from "lib/drupal";
+
+import { drupal } from "@/lib/drupal/drupal-client";
 
 export default async function handler(
   req: NextApiRequest,
